feat: add solution of challenge 3
Althought it's mentioned not to compute the answer using code, I still write
the relevant helper function needed to compute the byte by which each char
has been encrypted. Fetch the distribution of characters and xor'ed each
char with others to get new distribution of encrypted data chars. the input
cipher data is then compared with the new distribuion of encrypted data chars
and the one with maximum resemblence is picked.

Signed-off-by: 0xKanekiKen <100861945+0xKanekiKen@users.noreply.github.com>